\name{get.so}
\alias{get.so}
\title{Get code block from a SO question}
\usage{
  get.so(id, index = NULL, cat.output = FALSE)
}
\arguments{
  \item{id}{question id}

  \item{index}{index of the code block to return (one
  number only)}

  \item{cat.output}{cat the content of the code blocks}
}
\description{
  Get code block from a SO question
}

