\name{tmp.install.packages}
\alias{tmp.install.packages}
\title{Install a package and its dependencies just for the current session}
\usage{
  tmp.install.packages(pack, dependencies = TRUE, ...)
}
\arguments{
  \item{pack}{package name}

  \item{dependencies}{dependencies argument passed to
  install.packages}

  \item{...}{further arguments passed to
  install.packages()}
}
\description{
  Install a package and its dependencies just for the
  current session
}

